20201106提交

+++ ledao-system/src/main/java/com/ledao/system/mapper/SysJudicialMapper.java
@@ -17,7 +17,7 @@ public interface SysJudicialMapper {
      * @param id 司法拍卖项目ID
      * @return 司法拍卖项目
      */
-    public SysJudicial selectSysJudicialById(Integer id);
+    public SysJudicial selectSysJudicialById(Long id);
 
     /**
      * 查询司法拍卖项目列表
@@ -49,7 +49,7 @@ public interface SysJudicialMapper {
      * @param id 司法拍卖项目ID
      * @return 结果
      */
-    public int deleteSysJudicialById(Integer id);
+    public int deleteSysJudicialById(Long id);
 
     /**
      * 批量删除司法拍卖项目
